backend/requirements.txt
# ===========================================
# Kalag Backend Dependencies
# ===========================================

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0  # PostgreSQL async driver
aiosqlite==0.19.0  # SQLite async driver (for local dev)
alembic==1.13.1  # Database migrations

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
slowapi==0.1.9  # Rate limiting
secure==0.3.0  # Security headers

# AI/LLM
google-generativeai==0.3.2
# google-genai is the newer SDK

# Vector Database
qdrant-client==1.7.0

# Document Processing
llama-parse==0.3.0
pdf2image==1.16.3
Pillow==10.2.0
pypdf==3.17.4

# Utilities
httpx==0.26.0
python-dotenv==1.0.0
tenacity==8.2.3  # Retry logic

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
